# Support copying 1-dim array to ideep