--- a/chainer/link.py
+++ b/chainer/link.py
@@ -417,6 +417,7 @@ Assign a Parameter object directly to an attribute within a \
             if isinstance(value, cuda.ndarray):
                 value = value.get()  # to numpy.ndarray
             if (isinstance(value, numpy.ndarray) and value.ndim in (1, 2, 4)):
+                # TODO(kmaehashi): Remove ndim validation once iDeep has fixed.
                 # Currently iDeep only supports (1, 2, 4)-dim arrays.
                 # Note that array returned from `ideep.array` may not be an
                 # iDeep mdarray, e.g., when the dtype is not float32.