--- a/tests/chainer_tests/test_variable.py
+++ b/tests/chainer_tests/test_variable.py
@@ -1714,7 +1714,7 @@ class TestLossScale(unittest.TestCase):
 
 
 @testing.parameterize(*testing.product({
-    'shape': [(3, 2), (2, 3, 4, 3)],
+    'shape': [(0,), (1,), (3, 2), (2, 3, 4, 3)],
     'dtype': [
         np.int8, np.int16, np.int32, np.int64, np.uint8, np.uint16, np.uint32,
         np.uint64, np.float16, np.float32, np.float64],
@@ -1791,7 +1791,7 @@ class TestIntel64(unittest.TestCase):
 
 
 @testing.parameterize(*testing.product({
-    'shape': [(), (0,), (3, 2, 3), (4, 4, 3, 2, 3)],
+    'shape': [(), (3, 2, 3), (4, 4, 3, 2, 3)],
     'dtype': [
         np.int8, np.int16, np.int32, np.int64,
         np.uint8, np.uint16, np.uint32, np.uint64,