Currently `to_intel64` does not copy 1-dim arrays to iDeep.
As a result, bias parameter of the Linear layer is not copied to iDeep.
This causes memory copy overhead every time Linear layer is evaluated.

`mnist.mnist.MLPPredict.time_predict` benchmark results over this PR is as follows:
https://github.com/chainer/chainer-benchmark/blob/master/benchmarks/mnist/mnist.py#L92

Before applying this PR:

```
=========== ========== ============
--             units / batchsize   
----------- -----------------------
  backend    1000 / 1   1000 / 100 
=========== ========== ============
 cpu-ideep   971.01μs     1.40ms   
=========== ========== ============
```

After applying this PR:

```
=========== ========== ============
--             units / batchsize   
----------- -----------------------
  backend    1000 / 1   1000 / 100 
=========== ========== ============
 cpu-ideep   830.86μs     1.29ms   
=========== ========== ============
```