@kmaehashi 2018-05-15
Thanks for comments, fixed accordingly.