@niboshi 2018-05-17
Please fix `test_variable.py` as well.