@kmaehashi 2018-05-21
Fixed tests.